Update Permutations_13
This Java program generates and displays all permutations of numbers from 1 to a user-given number (between 1 and 13). Using a recursive algorithm, it ensures that all possible permutations of these numbers are created.

How it works:

The user enters a number n (1 ≤ n ≤ 13).

The program creates a list of numbers from 1 to n.

The generatePermutations function uses recursion to generate and display all possible permutations.